feat: вынесена функция обработки этапа вынесена отдельно
